Adding a new file collection class for Merritt so that I can track digests for use with zenodo uploads
